Added CMVContainer.cxx, fixed missing links and includes
