+++ Detectors/TPC/calibration/src/TPCCalibrationLinkDef.h
@@ -123,4 +123,6 @@
 #pragma link C++ class o2::tpc::DigitAdd + ;
 #pragma link C++ class std::vector < o2::tpc::DigitAdd> + ;
 #pragma link C++ class o2::tpc::PressureTemperatureHelper + ;
+#pragma link C++ class o2::tpc::CMVContainer + ;
+
 #endif
